+++ packages/pluggableWidgets/custom-chart-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- We fixed an issue where static data and source attribute wouldn't merge properly.
+
 ## [1.2.3] - 2025-10-10
 
 ### Changed